refactor(fbw): create NetworkTile component
